requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
openai>=1.0.0
tiktoken>=0.5.0
